// src/app/api/contribuyente/models/contacto-contribuyente-lista-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ContactoContribuyenteListaResponse {
  activo?: number;
  claMedAbreviatura?: string;
  claMedContactoId?: number;
  claMedDescripcion?: string;
  contribuyenteNumero?: number;
  descripcion?: string;
  estadoDsc?: string;
  medConContribuyenteId?: number;
  municipalidadId?: number;
  principal?: number;
  tipMedAbreviatura?: string;
  tipMedContactoId?: number;
  tipMedDescripcion?: string;
}